+++ hummingbot/strategy/uniswap_v3_lp/start.py
@@ -1,8 +1,8 @@
 from decimal import Decimal
 
 from hummingbot.strategy.market_trading_pair_tuple import MarketTradingPairTuple
-from hummingbot.strategy.uniswap_v3_lp.uniswap_v3_lp import UniswapV3LpStrategy
-from hummingbot.strategy.uniswap_v3_lp.uniswap_v3_lp_config_map import uniswap_v3_lp_config_map as c_map
+from hummingbot.strategy.amm_v3_lp.amm_v3_lp import UniswapV3LpStrategy
+from hummingbot.strategy.amm_v3_lp.amm_v3_lp_config_map import amm_v3_lp_config_map as c_map
 
 
 def start(self):
