+++ test/hummingbot/strategy/uniswap_v3_lp/test_uniswap_v3_lp_start.py
@@ -2,9 +2,9 @@ import unittest.mock
 from decimal import Decimal
 from test.hummingbot.strategy import assign_config_default
 
-import hummingbot.strategy.uniswap_v3_lp.start as uniswap_v3_lp_start
-from hummingbot.strategy.uniswap_v3_lp.uniswap_v3_lp import UniswapV3LpStrategy
-from hummingbot.strategy.uniswap_v3_lp.uniswap_v3_lp_config_map import uniswap_v3_lp_config_map
+import hummingbot.strategy.amm_v3_lp.start as amm_v3_lp_start
+from hummingbot.strategy.amm_v3_lp.amm_v3_lp import UniswapV3LpStrategy
+from hummingbot.strategy.amm_v3_lp.amm_v3_lp_config_map import amm_v3_lp_config_map
 
 
 class UniswapV3LpStartTest(unittest.TestCase):
@@ -15,14 +15,14 @@ class UniswapV3LpStartTest(unittest.TestCase):
         self.markets = {"uniswapLP": None}
         self.notifications = []
         self.log_errors = []
-        assign_config_default(uniswap_v3_lp_config_map)
-        uniswap_v3_lp_config_map.get("strategy").value = "uniswap_v3_lp"
-        uniswap_v3_lp_config_map.get("connector").value = "uniswapLP"
-        uniswap_v3_lp_config_map.get("market").value = "ETH-USDT"
-        uniswap_v3_lp_config_map.get("fee_tier").value = "LOW"
-        uniswap_v3_lp_config_map.get("price_spread").value = Decimal("1")
-        uniswap_v3_lp_config_map.get("amount").value = Decimal("1")
-        uniswap_v3_lp_config_map.get("min_profitability").value = Decimal("10")
+        assign_config_default(amm_v3_lp_config_map)
+        amm_v3_lp_config_map.get("strategy").value = "amm_v3_lp"
+        amm_v3_lp_config_map.get("connector").value = "uniswapLP"
+        amm_v3_lp_config_map.get("market").value = "ETH-USDT"
+        amm_v3_lp_config_map.get("fee_tier").value = "LOW"
+        amm_v3_lp_config_map.get("price_spread").value = Decimal("1")
+        amm_v3_lp_config_map.get("amount").value = Decimal("1")
+        amm_v3_lp_config_map.get("min_profitability").value = Decimal("10")
 
     def _initialize_market_assets(self, market, trading_pairs):
         pass
@@ -39,8 +39,8 @@ class UniswapV3LpStartTest(unittest.TestCase):
     def error(self, message, exc_info):
         self.log_errors.append(message)
 
-    @unittest.mock.patch('hummingbot.strategy.uniswap_v3_lp.uniswap_v3_lp.UniswapV3LpStrategy.add_markets')
-    def test_uniswap_v3_lp_strategy_creation(self, mock):
-        uniswap_v3_lp_start.start(self)
+    @unittest.mock.patch('hummingbot.strategy.amm_v3_lp.amm_v3_lp.UniswapV3LpStrategy.add_markets')
+    def test_amm_v3_lp_strategy_creation(self, mock):
+        amm_v3_lp_start.start(self)
         self.assertEqual(self.strategy._amount, Decimal(1))
         self.assertEqual(self.strategy._min_profitability, Decimal("10"))
